requirements.txt
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
Pillow==10.1.0
torch==2.1.0
ultralytics==8.0.222
